backend/requirements-lakebase.txt
# Web Democracy - Databricks Lakebase Requirements
# Lakebase è PostgreSQL-compatibile, quindi usiamo le stesse dipendenze

# Dipendenze base (identiche a requirements.txt)
fastapi==0.115.0
uvicorn[standard]==0.32.1
psycopg2-binary==2.9.10
SQLAlchemy==2.0.35
pydantic==2.10.2
python-multipart==0.0.12
python-dotenv==1.0.1

# Nota: Lakebase usa il protocollo PostgreSQL (porta 5432)
# Non serve databricks-sql-connector che è per SQL Warehouse (porta 443)

